+++ src/dudu/project_driver.hpp
@@ -4,6 +4,7 @@
 
 namespace dudu {
 
+std::filesystem::path clean_project(const std::filesystem::path& dir);
 void init_project(const std::filesystem::path& dir);
 void new_project(const std::filesystem::path& dir);